+++ CHANGELOG.md
@@ -18,3 +18,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implemented the Display component 
 - Implemented the Button component 
 - Implemented the ButtonPanel component
+- Flex display mode
+- Styled the Display component
+- Styled the ButtonPanel component
+- Styled the Button component
+- Passed "color" and "wide" props to the Button
+
+### Fixed
+- `calcRow` height to be 100px